API_frontend_flask/requirements_frontend.txt
Flask==3.0.0
Jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0 